section: add important info section that directs to terms and conditions
