+++ src/pages/Communities/Communities.jsx
@@ -1,90 +1,11 @@
-import PropTypes from 'prop-types';
-
-import Container from '../../components/Container';
-import { Header } from '../../components/Header/Header';
-import styles from './Communities.module.css';
-
-import help from '../../assets/communities/help.svg';
-import community from '../../assets/communities/community.svg';
-import objective from '../../assets/communities/objective.svg';
-import accountability from '../../assets/communities/accountability.svg';
-import sustainability from '../../assets/communities/sustainability.svg';
-import geographicFocus from '../../assets/communities/geagraphic-focus.svg';
-
-const cardData = [
-  {
-    image: help,
-    title: 'Demonstrated Need',
-    description:
-      'Communities must provide evidence of significant educational barriers, such as limited access to technology, or lack of essential learning resources.'
-  },
-  {
-    image: community,
-    title: 'Community Commitment',
-    description:
-      'Active participation from local leadership, parents, and community members is essential for successful implementation and sustainability of programmes.'
-  },
-  {
-    image: objective,
-    title: 'Clear Educational Objectives',
-    description:
-      'Applications must outline specific, measurable educational goals that the Edustipend support will help achieve within the community.'
-  },
-  {
-    image: accountability,
-    title: 'Accountability & Transparency',
-    description:
-      'Recipient communities must agree to transparent reporting on the use of funds/resources and allow for periodic impact assessments by Edustipend.'
-  },
-  {
-    image: sustainability,
-    title: 'Sustainability Plan',
-    description:
-      'Preference is given to communities that present a vision for maintaining the educational benefits beyond the initial support period.'
-  },
-  {
-    image: geographicFocus,
-    title: 'Geographic Focus',
-    description:
-      'Currently, our community support initiatives are focused within Nigeria. Specific regions may be prioritized based on strategic impact assessments.'
-  }
-];
-
-const CommunityCard = ({ image, title, description }) => (
-  <div className={styles.card}>
-    <img className={styles.cardImg} src={image} alt={title} />
-    <Header size="small" color="neutral" className={styles.cardTitle}>
-      {title}
-    </Header>
-    <p>{description}</p>
-  </div>
-);
+import Eligibility from '../../sections/Communities/Eligibility/Eligibility';
+import ImportantInformation from '../../sections/Communities/ImportantInformation/ImportantInformation';
 
 export const Communities = () => {
   return (
-    <section className={styles.bg}>
-      <Container className={styles.container}>
-        <article className={styles.contentHeader}>
-          <Header size="medium" color="neutral" className={styles.header}>
-            Community Eligibility Criteria
-          </Header>
-          <p className={styles.contentDescription}>
-            Our support is directed towards communities that demonstrate genuine need and a commitment to educational advancement.
-          </p>
-        </article>
-
-        <div className={styles.cardGrids}>
-          {cardData.map(({ image, title, description }) => (
-            <CommunityCard key={title} image={image} title={title} description={description} />
-          ))}
-        </div>
-      </Container>
-    </section>
+    <div>
+      <Eligibility />
+      <ImportantInformation />
+    </div>
   );
-};
-
-CommunityCard.propTypes = {
-  image: PropTypes.string.isRequired,
-  title: PropTypes.string.isRequired,
-  description: PropTypes.string.isRequired
 };
