--- src/pages/Communities/Communities.module.css
@@ -1,76 +1,0 @@
-.container {
-  padding-block-start: 183px;
-  padding-block-end: 50px;
-  background-color: rgba(88, 1, 255, 0.05);
-  border-radius: 8px;
-}
-
-.contentHeader {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  /* background-color: orange; */
-  margin-block-end: 40px;
-}
-
-.header,
-.contentDescription {
-  text-align: center;
-}
-
-.header {
-  font-size: 36px;
-  color: rgba(31, 41, 55, 1);
-}
-
-.contentDescription {
-  font-size: 18px;
-  font-family: 'Manrope', sans-serif;
-  font-weight: 400;
-}
-
-.cardGrids {
-  display: grid;
-  grid-template-columns: 1fr; /* default for mobile */
-  gap: 1.5rem;
-  padding: 1rem;
-  /* background-color: orange; */
-}
-
-.cardTitle {
-  font-size: 20px;
-  font-weight: 700;
-  font-family: 'Manrope', sans-serif;
-}
-
-@media (min-width: 768px) {
-  .cardGrids {
-    grid-template-columns: repeat(2, 1fr); /* tablets */
-  }
-}
-
-@media (min-width: 1024px) {
-  .cardGrids {
-    grid-template-columns: repeat(3, 1fr); /* desktop */
-  }
-}
-
-.card {
-  background: white;
-  border-radius: 8px;
-  padding: 24px;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-  text-align: left;
-
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-
-  /* max-width: 362px; */
-}
-
-.cardImg {
-  width: 38px;
-  height: 34.75px;
-}
