+++ src/sections/Communities/Eligibility/constant.js
@@ -1,0 +1,50 @@
+import help from '../../../assets/communities/help.svg';
+import community from '../../../assets/communities/community.svg';
+import objective from '../../../assets/communities/objective.svg';
+import accountability from '../../../assets/communities/accountability.svg';
+import sustainability from '../../../assets/communities/sustainability.svg';
+import geographicFocus from '../../../assets/communities/geagraphic-focus.svg';
+
+export const cardData = [
+  {
+    image: help,
+    title: 'Demonstrated Need',
+    description:
+      'Communities must provide evidence of significant educational barriers, such as limited access to technology, or lack of essential learning resources.'
+  },
+  {
+    image: community,
+    title: 'Community Commitment',
+    description:
+      'Active participation from local leadership, parents, and community members is essential for successful implementation and sustainability of programmes.'
+  },
+  {
+    image: objective,
+    title: 'Clear Educational Objectives',
+    description:
+      'Applications must outline specific, measurable educational goals that the Edustipend support will help achieve within the community.'
+  },
+  {
+    image: accountability,
+    title: 'Accountability & Transparency',
+    description:
+      'Recipient communities must agree to transparent reporting on the use of funds/resources and allow for periodic impact assessments by Edustipend.'
+  },
+  {
+    image: sustainability,
+    title: 'Sustainability Plan',
+    description:
+      'Preference is given to communities that present a vision for maintaining the educational benefits beyond the initial support period.'
+  },
+  {
+    image: geographicFocus,
+    title: 'Geographic Focus',
+    description:
+      'Currently, our community support initiatives are focused within Nigeria. Specific regions may be prioritized based on strategic impact assessments.'
+  }
+];
+
+export const elgibilityCriteria = {
+  title: 'Community Eligibility Criteria',
+  description: 'Our support is directed towards communities that demonstrate genuine need and a commitment to educational advancement.'
+};
