+++ src/sections/Communities/ImportantInformation/ImportantInfromation.module.css
@@ -1,0 +1,44 @@
+.bg {
+  margin-block: 64px;
+}
+
+.content {
+  box-shadow: 0px 4px 6px -4px rgba(0, 0, 0, 0.1);
+  box-shadow: 0px 10px 15px -3px rgba(0, 0, 0, 0.1);
+  background-color: rgba(243, 244, 246, 1);
+
+  padding-block: 40px;
+  padding-inline: 16px;
+}
+
+.title {
+  color: rgba(31, 41, 55, 1);
+  font-size: 30px;
+  line-height: 36px;
+
+  margin-block-end: 16px;
+}
+
+.title,
+.text {
+  text-align: center;
+}
+
+.text {
+  font-family: 'Manrope', sans-serif;
+  font-size: 18px;
+  font-weight: 400;
+  line-height: 28px;
+
+  margin-block-end: 46px;
+}
+
+.infoButton {
+  width: fit-content;
+}
+
+.buttonContainer {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
