+++ src/sections/Communities/ImportantInformation/constant.js
@@ -1,0 +1,4 @@
+export const importantInfo = {
+  title: 'Important Information',
+  text: 'Before applying or participating, please ensure you have read and understood our policies.'
+};
